Separate some of the code from PartsOfPlanTable into a separate component - PartsOfPlanElements.tsx to make it leaner.
